requirements.txt
python-telegram-bot==21.9
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
anthropic==0.40.0
python-dotenv==1.0.1